#14 Refactor code to find inital thresholds out to a separate class that will be easier to unit test.
